requirements.txt
python-magic
ssdeep
yara-python
requests
paramiko
pexpect